requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
astor=0.8.1=pypi_0
astunparse=1.6.3=pypi_0
autograd=1.3=pypi_0
blas=1.0=mkl
blosc=1.20.1=hd408876_0
brotli=1.0.9=he6710b0_2
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.1.19=h06a4308_1
cached-property=1.5.2=pypi_0
cachetools=4.2.1=pypi_0
certifi=2020.12.5=py36h06a4308_0
chardet=4.0.0=pypi_0
charls=2.1.0=he6710b0_2
cloudpickle=1.6.0=py_0
cudatoolkit=10.1.243=h6bb024c_0
cvxopt=1.2.0=py36hfa32c7d_0
cvxpy=1.1.7=pypi_0
cycler=0.10.0=py36_0
cytoolz=0.11.0=py36h7b6447c_0
dask-core=2020.12.0=pyhd3eb1b0_0
decorator=4.4.2=py_0
docopt=0.6.2=pypi_0
ecos=2.0.7.post1=pypi_0
flatbuffers=1.12=pypi_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=pypi_0
gast=0.3.3=pypi_0
giflib=5.1.4=h14c3975_1
glpk=4.65=h3ceedfd_2
gmp=6.1.2=h6c8ec71_1
google-auth=1.28.0=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.32.0=pypi_0
gsl=2.4=h14c3975_4
h5py=2.10.0=pypi_0
idna=2.10=pypi_0
imagecodecs=2020.5.30=py36hfa7d478_2
imageio=2.9.0=py_0
importlib-metadata=3.10.0=pypi_0
intel-openmp=2020.2=254
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jxrlib=1.1=h7b6447c_2
keras=2.2.4=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.0=py36h2531618_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libaec=1.0.4=he6710b0_1
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libwebp=1.0.1=h8e7db2f_0
libzopfli=1.0.3=he6710b0_0
lz4-c=1.9.2=heb0550a_3
markdown=3.3.4=pypi_0
matplotlib-base=3.3.2=py36h817c723_0
matplotlib-label-lines=0.3.8=pypi_0
metis=5.1.0=hf484d3e_4
mkl=2020.2=256
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.2.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
mock=4.0.3=pypi_0
ncurses=6.2=he6710b0_1
networkx=2.5=py_0
ninja=1.10.2=py36hff7bd54_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py36_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h27cfd23_0
opt-einsum=3.3.0=pypi_0
osqp=0.6.1=pypi_0
pandas=0.24.2=py36he6710b0_0
pillow=8.2.0=py36he98fc37_0
pip=20.3.3=py36h06a4308_0
pipreqs=0.4.10=pypi_0
protobuf=3.14.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=2.4.7=py_0
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=py_0
pytorch=1.6.0=py3.6_cuda10.1.243_cudnn7.6.3_0
pytz=2020.4=pyhd3eb1b0_0
pywavelets=1.1.1=py36h7b6447c_2
pyyaml=5.3.1=py36h7b6447c_1
readline=8.0=h7b6447c_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-image=0.17.2=py36hdf5156a_0
scikit-learn=0.23.2=py36h0573a6f_0
scipy=1.5.2=py36h0b6359f_0
scs=2.1.2=pypi_0
setuptools=51.0.0=py36h06a4308_2
six=1.15.0=py36h06a4308_0
snappy=1.1.8=he6710b0_0
sqlite=3.33.0=h62c20be_0
suitesparse=5.2.0=h9e4a6bb_0
tarjan=0.2.3.2=pypi_0
tbb=2020.3=hfd86e86_0
tensorboard=2.4.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorboardx=2.1=pypi_0
tensorflow-estimator=2.4.0=pypi_0
tensorflow-gpu=2.4.1=pypi_0
termcolor=1.1.0=pypi_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.12.8=pyhd3eb1b0_0
tk=8.6.10=hbc83047_0
toolz=0.11.1=py_0
torchvision=0.7.0=py36_cu101
tornado=6.1=py36h27cfd23_0
tqdm=4.56.0=pyhd3eb1b0_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.26.3=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarg=0.1.9=pypi_0
zipp=3.4.1=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
